+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.2] - 2020-09-04
 ### Added
  - Added HarmonyLang extension icon and Harmony file icons. VS Code currently does not support icon fallback, so the latter is currently disabled.
+ - Added `Run Harmony File` command.
+ - Added Harmony Output window.
 
 ## [0.0.1] - 2020-09-04
 ### Added